refactor(evaluation): consolidate blocked-state submodules into single file
state/blocked/{branch_sync,ci,review,generic}.rs はそれぞれ 1 つの enum を持つ
6〜10 行のファイルだった。4 階層パス `state/blocked/ci.rs` に対して中身が小さ
すぎ、import 文も `state::blocked::ci::CiState` の冗長な形になっていた。

4 つの enum を `state/blocked.rs` の `BlockedState` 構造体と同じファイルに統合。
import は `state::blocked::{BlockedState, CiState, ...}` のフラットな形になる。

Refs: #336
