--- src/contexts/evaluation/domain/prompt/pull_request/state/blocked/ci.rs
@@ -1,10 +1,0 @@
-/// CI のブロッカー評価状態
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum CiState {
-    /// チェックが失敗またはキャンセルされている
-    Fail,
-    /// 手動アクションが必要なチェックが存在する
-    ActionRequired,
-    /// チェックが実行中
-    Pending,
-}
