--- src/contexts/evaluation/domain/prompt/pull_request/state/blocked/generic.rs
@@ -1,6 +1,0 @@
-/// 汎用ブロッカー評価状態
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum GenericBlockedState {
-    /// API で原因を特定できないブロック（`mergeStateStatus == "BLOCKED"` かつ他シグナルすべて None）
-    BlockedUnknown,
-}
