--- src/contexts/evaluation/domain/prompt/pull_request/state/blocked/review.rs
@@ -1,8 +1,0 @@
-/// レビューのブロッカー評価状態
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum ReviewState {
-    /// レビュアーが変更を要求している
-    ChangesRequested,
-    /// レビュアーがまだアサインされていない
-    ReviewRequired,
-}
